moving jointDemand code - fix compile errors
