+++ contribs/commercialTrafficApplications/pom.xml
@@ -9,4 +9,19 @@
   <artifactId>commercialTrafficApplications</artifactId>
   <name>commercialTrafficApplications</name>
 
+    <dependencies>
+
+        <dependency>
+            <groupId>org.matsim.contrib</groupId>
+            <artifactId>freight</artifactId>
+            <version>12.0-SNAPSHOT</version>
+        </dependency>
+
+        <dependency>
+            <groupId>org.matsim.contrib</groupId>
+            <artifactId>drt</artifactId>
+            <version>12.0-SNAPSHOT</version>
+        </dependency>
+
+    </dependencies>
 </project>
