+++ contribs/commercialTrafficApplications/src/main/java/org/matsim/contrib/commercialTrafficApplications/jointDemand/commercialJob/CommercialJobScoreCalculator.java
@@ -1,4 +1,4 @@
-package commercialtraffic.commercialJob;
+package org.matsim.contrib.commercialTrafficApplications.jointDemand.commercialJob;
 
 interface CommercialJobScoreCalculator {
