+++ contribs/commercialTrafficApplications/src/main/java/org/matsim/contrib/commercialTrafficApplications/jointDemand/commercialJob/CommercialTrafficConfigGroup.java
@@ -17,7 +17,7 @@
  *                                                                         *
  * *********************************************************************** */
 
-package commercialtraffic.commercialJob;/*
+package org.matsim.contrib.commercialTrafficApplications.jointDemand.commercialJob;/*
  * created by jbischoff, 08.05.2019
  */
