+++ contribs/commercialTrafficApplications/src/test/java/org/matsim/contrib/commercialTrafficApplications/jointDemand/TestScenarioGeneration.java
@@ -18,10 +18,8 @@
  *                                                                         *
  * *********************************************************************** */
 
-package commercialtraffic;
+package org.matsim.contrib.commercialTrafficApplications.jointDemand;
 
-import commercialtraffic.commercialJob.CommercialJobUtils;
-import commercialtraffic.commercialJob.CommercialTrafficConfigGroup;
 import org.matsim.api.core.v01.Coord;
 import org.matsim.api.core.v01.Id;
 import org.matsim.api.core.v01.Scenario;
@@ -30,6 +28,8 @@ import org.matsim.api.core.v01.network.Link;
 import org.matsim.api.core.v01.population.Activity;
 import org.matsim.api.core.v01.population.Person;
 import org.matsim.api.core.v01.population.Plan;
+import org.matsim.contrib.commercialTrafficApplications.jointDemand.commercialJob.CommercialJobUtils;
+import org.matsim.contrib.commercialTrafficApplications.jointDemand.commercialJob.CommercialTrafficConfigGroup;
 import org.matsim.contrib.freight.carrier.*;
 import org.matsim.core.config.Config;
 import org.matsim.core.network.io.MatsimNetworkReader;
